Update patch set 4
Patch Set 4: Code-Review-1

There is a delay/missing recreate calls happening when I set % to inside or hidden. Sometimes it works others I need to tab statusbar for it to update. Minor bug, if we can nip it this week great, else we can do next (add it to redmine)

Patch-set: 4