Update patch set 2
Patch Set 2:

There is a delay/missing recreate calls happening when I set % to inside or hidden when the icon is set to landscape or portrait. Sometimes it works others I need to tap statusbar for it to update. Minor bug, if we can nip it this week great, else we can do next (add it to redmine.

Patch-set: 2
Label: Code-Review=0