Update patch set 1
Patch Set 1:

> I don't think this is required.
 > When using the logical or and the first statement is true, the
 > second statement shouldn't get checked because the result is
 > already known.
 > Thus, source != null should always be true when evaluated.

You are right, but the error came from there. I wonder if it's because of the async stuff and the threadpool executor, doing parallel operations. So it checked if source was !=null, but before doing the next .isRecycled check source became null. It's the only cause i can think about.

Patch-set: 1