Update patch set 1
Patch Set 1:

I haven't had the FC yet

 > > You are right, but the error came from there. I wonder if it's
 > > because of the async stuff and the threadpool executor, doing
 > > parallel operations. So it checked if source was !=null, but
 > before
 > > doing the next .isRecycled check source became null. It's the
 > only
 > > cause i can think about.
 > 
 > Hm, is there a way to reproduce this?

I haven't had the FC yet. Well, checking better now the whole thing, the user who reported the log had the FC while using one hand mode, and Slim doesn't have it. So it's probably due to this: 
https://github.com/ezio84/abc-n_frameworks_base/blob/8f176690a7de091da19b7992918ccefc302f2aeb/packages/SystemUI/src/com/android/systemui/slimrecent/RecentPanelView.java#L1321 when we get a cropped thumb for the one hand mode and recycle source (considering as said the threadpool doing parallel operations on the source bitmap).

In a nutshell, if i'm not wrong, this commit is probably useful only for roms with one hand mode and my stuff to crop the thumb (i could remove the source.recycle line but it's useful to free up the memory, and now we have this fix so...).

Patch-set: 1